feat: allow multiple vendors per framework
This patch updates the configuration to add the laminas-api-tools, mezzio, and zfcampus vendors.
Since these are subprojects of Laminas and Zend Framework, I needed to allow grouping them together for purposes of statistics and reporting.
As such, this patch also provides functionality to allow merging multiple vendors in order to report them under a single framework name.

When run, the landing page continues to report Laminas and Zend, but clicking through to their package pages, you will now see all qualifying packages across all vendor namespaces for those frameworks.
Since package links use the full package name to resolve, all package links continue to resolve correctly.
Moreover, since the Laminas and ZF packages use package prefixes, they are easily recognizable.

To accomplish the merging operations, I moved the calculation of the last year trend percentage to the constructor of `VendorData`.
This allowed me to remove any logic duplication, as the value is derived from other values passed to the constructor.

One potential improvement is to include the full package name, including vendor prefix, on the packages page; I'm willing to make the changes to the templates if this is desired.